Move custom hooks to their own files, add comments
